+++ nexeruncator/commands/extractjs/extractjs_test.go
@@ -1,0 +1,1 @@
+package extractjs_test