+++ nexeruncator/commands/insertjs/insertjs_test.go
@@ -1,0 +1,1 @@
+package insertjs_test