+++ CHANGELOG.md
@@ -2,13 +2,30 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.2.0] - 2022-12-27
+
+### Added
+
+- urfave v2 implemented for commandline parameters
+- commands added for extracting and inserting
+
+### Changed
+
+- not using flag package for commandline parameters
+- not using import anymore. using insert
+
+### Fixed
+
+- Links refers old repo are fixed
+
 ## [0.1.0] - 2022-12-23
 
 Initial release.
 
-nexeradicator can now extract javascript source from a nexe-compiled binary.
+nexeruncator can now extract javascript source from a nexe-compiled binary.
 It exports and and saved the javascript source as its original name.
-nexeradicator can also insert a new javascript source file into nexe-compiled
+nexeruncator can also insert a new javascript source file into nexe-compiled
 binary.
 
-[0.1.0]: https://github.com/blue-devil/nexeradicator/releases/tag/v0.1.0
+[0.2.0]: https://github.com/blue-devil/nexeruncator/compare/v0.1.0...v0.2.0
+[0.1.0]: https://github.com/blue-devil/nexeruncator/releases/tag/v0.1.0